# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
...
  models.AuthResponse:
    properties:
      mfa_challenge_id:
        type: string
      profile:
        $ref: '#/definitions/models.ProfileResponse'
      requires_mfa:
        type: boolean
      tokens:
        $ref: '#/definitions/models.TokenPair'
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
...
  models.MFABackupCodeRequest:
    properties:
      backup_code:
        type: string
      challenge_id:
        type: string
    required:
    - backup_code
    - challenge_id
    type: object
  models.MFADisableRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  models.MFAEnrollRequest:
    properties:
      type:
        enum:
        - TOTP
        - SMS
        - EMAIL
        type: string
    required:
    - type
    type: object
  models.MFAEnrollResponse:
    properties:
      backup_codes:
        items:
          type: string
        type: array
      factor_id:
        type: string
      qr_code_url:
        description: For TOTP
        type: string
      secret_key:
        description: For TOTP (backup)
        type: string
      type:
        type: string
    type: object
...
  models.MFAVerifyRequest:
    properties:
      code:
        type: string
      factor_id:
        type: string
    required:
    - code
    - factor_id
    type: object
  models.Photo:
    properties:
      height:
        type: integer
      mime_type:
        type: string
      name:
        type: string
      size:
        type: integer
      url:
        type: string
      width:
        type: integer
    type: object
...
  models.ProfileResponse:
    properties:
      avatar:
        $ref: '#/definitions/models.Photo'
      first_name:
        type: string
      id:
        type: string
      is_complete:
        type: boolean
      last_name:
        type: string
    type: object
...
  models.TokenPair:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      mfa_enabled:
        type: boolean
      phone_verified:
        type: boolean
    type: object
...
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@hamsaya.app
    name: API Support
    url: https://hamsaya.app/support
  description: A production-ready Go backend for a social media mobile application
    with posts, business profiles, marketplace, real-time chat, and location services.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://hamsaya.app/terms/
  title: Hamsaya Backend API
  version: "1.0"
paths:
...
  /mfa/backup-codes/count:
    get:
      description: Get the count of unused backup codes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  type: integer
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get backup codes count
      tags:
      - mfa
  /mfa/backup-codes/regenerate:
    post:
      description: Regenerate backup codes for MFA
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Regenerate backup codes
      tags:
      - mfa
  /mfa/disable:
    post:
      consumes:
      - application/json
      description: Disable multi-factor authentication for the user
      parameters:
      - description: MFA disable request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MFADisableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Disable MFA
      tags:
      - mfa
  /mfa/enroll:
    post:
      consumes:
      - application/json
      description: Enroll TOTP (Time-based One-Time Password) for MFA
      parameters:
      - description: MFA enrollment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MFAEnrollRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.MFAEnrollResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Enroll TOTP
      tags:
      - mfa
  /mfa/verify-backup-code:
    post:
      consumes:
      - application/json
      description: Verify MFA challenge with a backup code
      parameters:
      - description: Backup code request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MFABackupCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Verify with backup code
      tags:
      - mfa
  /mfa/verify-enrollment:
    post:
      consumes:
      - application/json
      description: Verify MFA enrollment with code from authenticator app
      parameters:
      - description: MFA verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MFAVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Verify MFA enrollment
      tags:
      - mfa
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"